pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = "这是我学习drf的一个项目"
authors = ["Andy Z Wright <zhangruibin021@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
PyMySQL = "^1.0.2"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
